requirements.txt
asgiref>=3.7.0
charset-normalizer>=2.0.0
cssselect2>=0.7.0
dj-database-url>=0.5.0
Django==4.2.16
django-classy-tags>=3.0.1
django-cms==3.11.4
django-filer>=2.2.4
django-formtools>=2.3
django-js-asset>=2.0.0
django-mptt>=0.14.0
django-polymorphic>=3.0.0
django-sekizai>=3.0.1
django-treebeard>=4.5.1
django-tinymce>=4.1.0
djangocms-admin-style>=3.1.1
djangocms-attributes-field>=2.0.0
djangocms-bootstrap4>=2.0.0
djangocms-file>=3.0.0,<5.0.0
djangocms-googlemap>=2.0.0,<4.0.0
djangocms-icon>=2.0.0
djangocms-link>=3.0.0,<5.0.0
djangocms-picture>=3.0.0,<5.0.0
djangocms-style>=3.0.0
djangocms-text-ckeditor>=5.0.0
djangocms-video>=3.0.0,<5.0.0
easy-thumbnails>=2.8.0
html5lib>=1.1
jsmin>=3.0.0
lxml>=4.9.0
pillow>=9.0.0
psycopg2-binary>=2.9.0
pytz>=2023.3
reportlab>=4.0.0
six>=1.16.0
sqlparse>=0.4.4
svglib>=1.5.1
tinycss2>=1.2.1
tzdata>=2023.3
tzlocal>=5.0
webencodings>=0.5.1
python-decouple
django-widget-tweaks
pandas>=2.0
openpyxl>=3.1
waitress>=3.0.0
django-csp>=3.8
whitenoise>=6.6.0
waitress>=3.0.0